Update index.html and all files of JavaScript
I created a function which when user click on add to cart button, this function will check that the product user wants to add ion cart is already presented in add to cart section or not. if it's presented then this function will run and this function will also create a notification div that it's elkectedproduct hads been added to add o cart section.